benchmark/tools/requirements.txt
matplotlib>=3.5.0
Pillow>=8.0.0
numpy>=1.20.0 